pyproject.toml
[project]
name = "saint-seiya"
version = "0.1.0"
description = ""
authors = [
    {name = "Meu Nome",email = "jeffer_sg@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy (==1.26.4)",
    "opencv-python (==4.7.0.72)",
    "easyocr (>=1.7.2,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "selenium (>=4.35.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
